+++ index.d.ts
@@ -132,6 +132,7 @@ export declare namespace be {
     function desktop(agent?: string): boolean;
 
     function sha1(value: string): boolean;
+    function sha256(value: string): boolean;
     function md5(value: string): boolean;
 
     function email(value: string): boolean;